requirements.txt
fastapi==0.109.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
secure

